FE - rename: 공통 컴포넌트 폴더 이동
- layout 폴더에서 components 폴더로 이동하였음.
