+++ client/src/components/common/Error.tsx
