+++ client/src/components/common/Header.tsx
